+++ SkinMetrolook.php
@@ -1,7 +1,6 @@
 <?php
 /**
- * Vector - Modern version of MonoBook with fresh look and many usability
- * improvements.
+ * Metrolook - Metro look for website.
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -23,7 +22,7 @@
  */
 
 /**
- * SkinTemplate class for Vector skin
+ * SkinTemplate class for Metrolook skin
  * @ingroup Skins
  */
 class SkinMetrolook extends SkinTemplate {